requirements.txt
flask>=3.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
cloudscraper>=1.2.71
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
